requirements.txt
Django~=4.0.3
qrcode~=7.3.1
Pillow~=9.1.0
psycopg2==2.9.3
python-decouple==3.6